fix: move resolve credential manifest from didcommwallet to vcwallet.
